Php + C + C++ + JavaScript + Html
The aim of Talksta is to provide a single platform for students and the staffs to interact
with each other. This website also provides the features of writing and posting a post or
any event all at one place. This project aims to develop a website which provides a
Communication among peoples on network, which works quite similar to Social Media
Site. Talksta project is designed only for students of the college. Students can connect by
sending friend requests. Students are also able to send messages to each other which is the
major feature of any social networking platform.